requirements.txt
django>=1.5
djangorestless 
django-cors-headers
pillow
stripe
requests
